refactor: consolidate managers into UIManager
