--- Gamblers Revenge/Assets/Scripts/LoseManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7e8be0e35009e4da68af7a041a0e1039
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
