+++ Gamblers Revenge/Assets/Scripts/PlayerController.cs
@@ -76,7 +76,7 @@ public class PlayerController : MonoBehaviour
         string[] optionNames = picks.Select(u => u.ToString()).ToArray();
 
         // 3) show UI
-        UpgradeManager.instance.ShowUpgradeScreen(optionNames, choice =>
+        UIManager.instance.ShowUpgradeScreen(optionNames, choice =>
         {
             // 4) apply the picked upgrade
             UpgradeEffects.UpgradeType chosen = picks[choice];
