--- Gamblers Revenge/Assets/Scripts/UpgradeManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f0941f5bbbf7547f687cd46e5f927986
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
